arreglo de tipo de datos a long

+++ src/main/java/com/pruebaSpring/demo/models/dao/IUsuariosDao.java
@@ -3,5 +3,5 @@ package com.pruebaSpring.demo.models.dao;
 import com.pruebaSpring.demo.models.entity.Usuarios;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IUsuariosDao extends JpaRepository<Usuarios, Integer> {
+public interface IUsuariosDao extends JpaRepository<Usuarios, Long> {
 }
